Update Architecture : Move file to folder
